fix: code updated, image picker finialized, env file deleted
