--- .env
@@ -1,1 +1,0 @@
-API_URL = "http://localhost:8080/"
