+++ .gitignore
@@ -34,3 +34,7 @@ yarn-error.*
 
 # typescript
 *.tsbuildinfo
+
+# dotenv environment variables file
+.env
+.env.*
